Move SmelteryFuelRecipe
